Added new creature
The creature in the Creature class is a small green square that moves similarly to the griddies. However, these creatures are prey! When the creatures collide with the griddies, they get weaker and weaker (lose energy) as if they are getting eaten. The griddies get more blue in color the more they are nourrished.